requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
authx==1.4.3
bcrypt==4.1.2
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.4
fastapi-users==15.0.3
fastjsonschema==2.21.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
makefun==1.16.0
nbformat==5.10.4
passlib==1.7.4
platformdirs==4.5.1
psycopg2-binary==2.9.11
pwdlib==0.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
referencing==0.37.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
